refactor: deprecate cache implementation and types
