+++ .changeset/forty-rice-visit.md
@@ -1,0 +1,5 @@
+---
+'@backstage/backend-common': patch
+---
+
+Move `cache` implementation and types to the `@backstage/backend-defaults` package.
