+++ .changeset/lemon-weeks-lay.md
@@ -1,0 +1,5 @@
+---
+'@backstage/backend-defaults': patch
+---
+
+Remove dependency with `@backstage/backend-commons` package.
