--- packages/backend-common/src/cache/types.ts
@@ -1,39 +1,0 @@
-/*
- * Copyright 2020 The Backstage Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import type {
-  CacheService,
-  CacheServiceSetOptions,
-  CacheServiceOptions,
-} from '@backstage/backend-plugin-api';
-
-/**
- * @public
- * @deprecated Use `CacheService` from the `@backstage/backend-plugin-api` package instead
- */
-export type CacheClient = CacheService;
-
-/**
- * @public
- * @deprecated Use `CacheServiceSetOptions` from the `@backstage/backend-plugin-api` package instead
- */
-export type CacheClientSetOptions = CacheServiceSetOptions;
-
-/**
- * @public
- * @deprecated Use `CacheServiceOptions` from the `@backstage/backend-plugin-api` package instead
- */
-export type CacheClientOptions = CacheServiceOptions;
