+++ packages/backend-defaults/package.json
@@ -88,7 +88,6 @@
     "@aws-sdk/credential-providers": "^3.350.0",
     "@aws-sdk/types": "^3.347.0",
     "@backstage/backend-app-api": "workspace:^",
-    "@backstage/backend-common": "workspace:^",
     "@backstage/backend-dev-utils": "workspace:^",
     "@backstage/backend-plugin-api": "workspace:^",
     "@backstage/config": "workspace:^",
