+++ packages/backend-defaults/src/entrypoints/scheduler/lib/PluginTaskSchedulerImpl.test.ts
@@ -14,8 +14,11 @@
  * limitations under the License.
  */
 
-import { getVoidLogger } from '@backstage/backend-common';
-import { TestDatabaseId, TestDatabases } from '@backstage/backend-test-utils';
+import {
+  TestDatabaseId,
+  TestDatabases,
+  mockServices,
+} from '@backstage/backend-test-utils';
 import { ConflictError, NotFoundError } from '@backstage/errors';
 import { Duration } from 'luxon';
 import { migrateBackendTasks } from '../database/migrateBackendTasks';
@@ -53,7 +56,7 @@ describe('PluginTaskManagerImpl', () => {
     await migrateBackendTasks(knex);
     const manager = new PluginTaskSchedulerImpl(
       async () => knex,
-      getVoidLogger(),
+      mockServices.logger.mock(),
     );
     return { knex, manager };
   }
