+++ packages/backend-defaults/src/index.ts
@@ -20,9 +20,4 @@
  * @packageDocumentation
  */
 
-// This import is here as a workaround for a cyclic dependency bug where
-// backend-common must be loaded before backend-app-api
-// TODO(Rugvip): Remove this once backend-common is no longer used by backend-app-api
-import '@backstage/backend-common';
-
 export { createBackend } from './CreateBackend';
